[Lang] Replace internal representation of Python-scope ti.Matrix with numpy arrays

+++ tests/python/test_custom_struct.py
@@ -314,11 +314,11 @@ def test_compound_type_implicit_cast():
     int_value = f2i_taichi_scope()
     assert type(int_value) == int and int_value == 6
     int_value = f2i_python_scope()
-    assert type(int_value) == int and int_value == 6
+    assert type(int_value) == np.int64 and int_value == 6
     float_value = i2f_taichi_scope()
     assert type(float_value) == float and float_value == approx(6.0, rel=1e-4)
     float_value = i2f_python_scope()
-    assert type(float_value) == float and float_value == approx(6.0, rel=1e-4)
+    assert type(float_value) == np.float64 and float_value == approx(6.0, rel=1e-4)
 
 
 @test_utils.test()
